requirements.txt
# Install: pip install -r requirements.txt
# Python 3.11+ recommended.

langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-tavily>=0.1.0
google-genai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
